==> 0_initiation.ipynb <==
<p>In this early phase of the project, we will : </p> 
<ul> 
    <li> prepare the required data   </li> 
    <li> create data storage repositories</li> 
</ul>

### Table of Contents

* [1. Create Data Lake](#section1)
* [2. Define Cities](#section2)
* [3. Define Hotels](#section3)

# Predefined Functions
from modules import Funct as F

# global params
bucket_name = 'kayak-project'

### 1. Create Data Lake <a class="anchor" id="section1"></a>

F.create_bucket(bucket_name)

### 2. Define Cities <a class="anchor" id="section2"></a>

# global params
cities_list = ["Mont Saint Michel", "St Malo", "Bayeux", "Le Havre", "Rouen", "Paris", "Amiens", "Lille", "Strasbourg", 
               "Chateau du Haut Koenigsbourg", "Colmar", "Eguisheim", "Besancon", "Dijon", "Annecy", "Grenoble", "Lyon", 
               "Gorges du Verdon", "Bormes les Mimosas", "Cassis", "Marseille", "Aix en Provence", "Avignon", "Uzes", 
               "Nimes", "Aigues Mortes", "Saintes Maries de la mer", "Collioure", "Carcassonne", "Ariege", "Toulouse", 
               "Montauban", "Biarritz", "Bayonne", "La Rochelle"]

# save the list of cities in a txt file
F.write_txt(cities_list, 'data/cities.txt', mode ='w')

# save the txt file in the data lake
F.upload_file_dl('data/cities.txt', bucket_name, 'cities.txt')

### 3. Define Hotels <a class="anchor" id="section3"></a>

url_mont_saint_michel = "https://www.booking.com/searchresults.fr.html?label=gen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB&sid=a33dd63e1835e1ec9af64c715c2bc273&aid=304142&src=searchresults&error_url=https%3A%2F%2Fwww.booking.com%2Fsearchresults.fr.html%3Flabel%3Dgen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB%3Bsid%3Da33dd63e1835e1ec9af64c715c2bc273%3Btmpl%3Dsearchresults%3Bac_click_type%3Db%3Bac_position%3D0%3Bclass_interval%3D1%3Bdest_id%3D-1456928%3Bdest_type%3Dcity%3Bdtdisc%3D0%3Bfrom_sf%3D1%3Bgroup_adults%3D2%3Bgroup_children%3D0%3Biata%3DPAR%3Binac%3D0%3Bindex_postcard%3D0%3Blabel_click%3Dundef%3Bno_rooms%3D1%3Boffset%3D0%3Bpostcard%3D0%3Braw_dest_type%3Dcity%3Broom1%3DA%252CA%3Bsb_price_type%3Dtotal%3Bsearch_selected%3D1%3Bshw_aparth%3D1%3Bslp_r_match%3D0%3Bsrc%3Dindex%3Bsrc_elem%3Dsb%3Bsrpvid%3D2f5479ecc1720166%3Bss%3DParis%252C%2520%25C3%258Ele-de-France%252C%2520France%3Bss_all%3D0%3Bss_raw%3Dparis%3Bssb%3Dempty%3Bsshis%3D0%3Btop_ufis%3D1%26%3B&ss=Le+Mont-Saint-Michel%2C+Basse-Normandie%2C+France&is_ski_area=&ssne=Paris&ssne_untouched=Paris&city=-1456928&checkin_year=&checkin_month=&checkout_year=&checkout_month=&group_adults=2&group_children=0&no_rooms=1&from_sf=1&ss_raw=Mont+saint+&ac_position=0&ac_langcode=fr&ac_click_type=b&dest_id=900039327&dest_type=city&place_id_lat=48.636&place_id_lon=-1.511&search_pageview_id=2f5479ecc1720166&search_selected=true&search_pageview_id=2f5479ecc1720166&ac_suggestion_list_length=5&ac_suggestion_theme_list_length=0&nflt=ht_id%3D204"
url_saint_malo = "https://www.booking.com/searchresults.fr.html?label=gen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB&sid=a33dd63e1835e1ec9af64c715c2bc273&aid=304142&src=searchresults&error_url=https%3A%2F%2Fwww.booking.com%2Fsearchresults.fr.html%3Faid%3D304142%3Blabel%3Dgen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB%3Bsid%3Da33dd63e1835e1ec9af64c715c2bc273%3Btmpl%3Dsearchresults%3Bac_click_type%3Db%3Bac_position%3D0%3Bcity%3D-1456928%3Bclass_interval%3D1%3Bdest_id%3D900039327%3Bdest_type%3Dcity%3Bdtdisc%3D0%3Bfrom_sf%3D1%3Bgroup_adults%3D2%3Bgroup_children%3D0%3Binac%3D0%3Bindex_postcard%3D0%3Blabel_click%3Dundef%3Bno_rooms%3D1%3Boffset%3D0%3Bpostcard%3D0%3Braw_dest_type%3Dcity%3Broom1%3DA%252CA%3Bsb_price_type%3Dtotal%3Bsearch_selected%3D1%3Bshw_aparth%3D1%3Bslp_r_match%3D0%3Bsrc%3Dsearchresults%3Bsrc_elem%3Dsb%3Bsrpvid%3D83f67c62a2050151%3Bss%3DLe%2520Mont-Saint-Michel%252C%2520Basse-Normandie%252C%2520France%3Bss_all%3D0%3Bss_raw%3DMont%2520saint%3Bssb%3Dempty%3Bsshis%3D0%3Bssne%3DParis%3Bssne_untouched%3DParis%3Btop_ufis%3D1%26%3B&ss=Saint-Malo%2C+Bretagne%2C+France&is_ski_area=&ssne=Le+Mont-Saint-Michel&ssne_untouched=Le+Mont-Saint-Michel&city=900039327&checkin_year=&checkin_month=&checkout_year=&checkout_month=&group_adults=2&group_children=0&no_rooms=1&from_sf=1&ss_raw=Saint+ma&ac_position=0&ac_langcode=fr&ac_click_type=b&dest_id=-1466824&dest_type=city&place_id_lat=48.649616&place_id_lon=-2.025677&search_pageview_id=83f67c62a2050151&search_selected=true&search_pageview_id=83f67c62a2050151&ac_suggestion_list_length=5&ac_suggestion_theme_list_length=0&nflt=ht_id%3D204"
url_bayeux = "https://www.booking.com/searchresults.fr.html?label=gen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB&sid=a33dd63e1835e1ec9af64c715c2bc273&aid=304142&src=searchresults&error_url=https%3A%2F%2Fwww.booking.com%2Fsearchresults.fr.html%3Faid%3D304142%3Blabel%3Dgen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB%3Bsid%3Da33dd63e1835e1ec9af64c715c2bc273%3Btmpl%3Dsearchresults%3Bac_click_type%3Db%3Bac_position%3D0%3Bcity%3D900039327%3Bclass_interval%3D1%3Bdest_id%3D-1466824%3Bdest_type%3Dcity%3Bdtdisc%3D0%3Bfrom_sf%3D1%3Bgroup_adults%3D2%3Bgroup_children%3D0%3Binac%3D0%3Bindex_postcard%3D0%3Blabel_click%3Dundef%3Bno_rooms%3D1%3Boffset%3D0%3Bpostcard%3D0%3Braw_dest_type%3Dcity%3Broom1%3DA%252CA%3Bsb_price_type%3Dtotal%3Bsearch_selected%3D1%3Bshw_aparth%3D1%3Bslp_r_match%3D0%3Bsrc%3Dsearchresults%3Bsrc_elem%3Dsb%3Bsrpvid%3D646d7c78fcf40040%3Bss%3DSaint-Malo%252C%2520Bretagne%252C%2520France%3Bss_all%3D0%3Bss_raw%3DSaint%2520ma%3Bssb%3Dempty%3Bsshis%3D0%3Bssne%3DLe%2520Mont-Saint-Michel%3Bssne_untouched%3DLe%2520Mont-Saint-Michel%3Btop_ufis%3D1%26%3B&ss=Bayeux%2C+Basse-Normandie%2C+France&is_ski_area=&ssne=Saint-Malo&ssne_untouched=Saint-Malo&city=-1466824&checkin_year=&checkin_month=&checkout_year=&checkout_month=&group_adults=2&group_children=0&no_rooms=1&from_sf=1&ss_raw=baye&ac_position=0&ac_langcode=fr&ac_click_type=b&dest_id=-1410836&dest_type=city&place_id_lat=49.27605&place_id_lon=-0.70229&search_pageview_id=646d7c78fcf40040&search_selected=true&search_pageview_id=646d7c78fcf40040&ac_suggestion_list_length=5&ac_suggestion_theme_list_length=0&nflt=ht_id%3D204"
url_le_havre = "https://www.booking.com/searchresults.fr.html?label=gen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB&sid=a33dd63e1835e1ec9af64c715c2bc273&aid=304142&src=searchresults&error_url=https%3A%2F%2Fwww.booking.com%2Fsearchresults.fr.html%3Faid%3D304142%3Blabel%3Dgen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB%3Bsid%3Da33dd63e1835e1ec9af64c715c2bc273%3Btmpl%3Dsearchresults%3Bac_click_type%3Db%3Bac_position%3D0%3Bcity%3D-1466824%3Bclass_interval%3D1%3Bdest_id%3D-1410836%3Bdest_type%3Dcity%3Bdtdisc%3D0%3Bfrom_sf%3D1%3Bgroup_adults%3D2%3Bgroup_children%3D0%3Binac%3D0%3Bindex_postcard%3D0%3Blabel_click%3Dundef%3Bno_rooms%3D1%3Boffset%3D0%3Bpostcard%3D0%3Braw_dest_type%3Dcity%3Broom1%3DA%252CA%3Bsb_price_type%3Dtotal%3Bsearch_selected%3D1%3Bshw_aparth%3D1%3Bslp_r_match%3D0%3Bsrc%3Dsearchresults%3Bsrc_elem%3Dsb%3Bsrpvid%3D39147c853c7d016d%3Bss%3DBayeux%252C%2520Basse-Normandie%252C%2520France%3Bss_all%3D0%3Bss_raw%3Dbaye%3Bssb%3Dempty%3Bsshis%3D0%3Bssne%3DSaint-Malo%3Bssne_untouched%3DSaint-Malo%3Btop_ufis%3D1%26%3B&ss=Le+Havre%2C+Haute-Normandie%2C+France&is_ski_area=&ssne=Bayeux&ssne_untouched=Bayeux&city=-1410836&checkin_year=&checkin_month=&checkout_year=&checkout_month=&group_adults=2&group_children=0&no_rooms=1&from_sf=1&ss_raw=le+havre&ac_position=0&ac_langcode=fr&ac_click_type=b&dest_id=-1441598&dest_type=city&place_id_lat=49.49489&place_id_lon=0.107723&search_pageview_id=39147c853c7d016d&search_selected=true&search_pageview_id=39147c853c7d016d&ac_suggestion_list_length=5&ac_suggestion_theme_list_length=0&nflt=ht_id%3D204"
url_rouen = "https://www.booking.com/searchresults.fr.html?label=gen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB&sid=a33dd63e1835e1ec9af64c715c2bc273&aid=304142&src=searchresults&error_url=https%3A%2F%2Fwww.booking.com%2Fsearchresults.fr.html%3Faid%3D304142%3Blabel%3Dgen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB%3Bsid%3Da33dd63e1835e1ec9af64c715c2bc273%3Btmpl%3Dsearchresults%3Bac_click_type%3Db%3Bac_position%3D0%3Bcity%3D-1410836%3Bclass_interval%3D1%3Bdest_id%3D-1441598%3Bdest_type%3Dcity%3Bdtdisc%3D0%3Bfrom_sf%3D1%3Bgroup_adults%3D2%3Bgroup_children%3D0%3Binac%3D0%3Bindex_postcard%3D0%3Blabel_click%3Dundef%3Bno_rooms%3D1%3Boffset%3D0%3Bpostcard%3D0%3Braw_dest_type%3Dcity%3Broom1%3DA%252CA%3Bsb_price_type%3Dtotal%3Bsearch_selected%3D1%3Bshw_aparth%3D1%3Bslp_r_match%3D0%3Bsrc%3Dsearchresults%3Bsrc_elem%3Dsb%3Bsrpvid%3D347b7ca2ff8501e6%3Bss%3DLe%2520Havre%252C%2520Haute-Normandie%252C%2520France%3Bss_all%3D0%3Bss_raw%3Dle%2520havre%3Bssb%3Dempty%3Bsshis%3D0%3Bssne%3DBayeux%3Bssne_untouched%3DBayeux%3Btop_ufis%3D1%26%3B&ss=Rouen%2C+Haute-Normandie%2C+France&is_ski_area=&ssne=Le+Havre&ssne_untouched=Le+Havre&city=-1441598&checkin_year=&checkin_month=&checkout_year=&checkout_month=&group_adults=2&group_children=0&no_rooms=1&from_sf=1&ss_raw=rouen&ac_position=0&ac_langcode=fr&ac_click_type=b&dest_id=-1462807&dest_type=city&place_id_lat=49.44046&place_id_lon=1.09397&search_pageview_id=347b7ca2ff8501e6&search_selected=true&search_pageview_id=347b7ca2ff8501e6&ac_suggestion_list_length=5&ac_suggestion_theme_list_length=0&nflt=ht_id%3D204"
url_paris = "https://www.booking.com/searchresults.fr.html?label=gen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB&sid=a33dd63e1835e1ec9af64c715c2bc273&aid=304142&src=searchresults&error_url=https%3A%2F%2Fwww.booking.com%2Fsearchresults.fr.html%3Faid%3D304142%3Blabel%3Dgen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB%3Bsid%3Da33dd63e1835e1ec9af64c715c2bc273%3Btmpl%3Dsearchresults%3Bac_click_type%3Db%3Bac_position%3D0%3Bcity%3D-1441598%3Bclass_interval%3D1%3Bdest_id%3D-1462807%3Bdest_type%3Dcity%3Bdtdisc%3D0%3Bfrom_sf%3D1%3Bgroup_adults%3D2%3Bgroup_children%3D0%3Binac%3D0%3Bindex_postcard%3D0%3Blabel_click%3Dundef%3Bno_rooms%3D1%3Boffset%3D0%3Bpostcard%3D0%3Braw_dest_type%3Dcity%3Broom1%3DA%252CA%3Bsb_price_type%3Dtotal%3Bsearch_selected%3D1%3Bshw_aparth%3D1%3Bslp_r_match%3D0%3Bsrc%3Dsearchresults%3Bsrc_elem%3Dsb%3Bsrpvid%3D29777caed9a90010%3Bss%3DRouen%252C%2520Haute-Normandie%252C%2520France%3Bss_all%3D0%3Bss_raw%3Drouen%3Bssb%3Dempty%3Bsshis%3D0%3Bssne%3DLe%2520Havre%3Bssne_untouched%3DLe%2520Havre%3Btop_ufis%3D1%26%3B&ss=Paris%2C+%C3%8Ele-de-France%2C+France&is_ski_area=&ssne=Rouen&ssne_untouched=Rouen&city=-1462807&checkin_year=&checkin_month=&checkout_year=&checkout_month=&group_adults=2&group_children=0&no_rooms=1&from_sf=1&ss_raw=paris&ac_position=0&ac_langcode=fr&ac_click_type=b&dest_id=-1456928&dest_type=city&iata=PAR&place_id_lat=48.85668&place_id_lon=2.351476&search_pageview_id=29777caed9a90010&search_selected=true&search_pageview_id=29777caed9a90010&ac_suggestion_list_length=5&ac_suggestion_theme_list_length=0&nflt=ht_id%3D204"
url_amiens = "https://www.booking.com/searchresults.fr.html?label=gen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB&sid=a33dd63e1835e1ec9af64c715c2bc273&aid=304142&src=searchresults&error_url=https%3A%2F%2Fwww.booking.com%2Fsearchresults.fr.html%3Faid%3D304142%3Blabel%3Dgen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB%3Bsid%3Da33dd63e1835e1ec9af64c715c2bc273%3Btmpl%3Dsearchresults%3Bac_click_type%3Db%3Bac_position%3D0%3Bcity%3D-1462807%3Bclass_interval%3D1%3Bdest_id%3D-1456928%3Bdest_type%3Dcity%3Bdtdisc%3D0%3Bfrom_sf%3D1%3Bgroup_adults%3D2%3Bgroup_children%3D0%3Biata%3DPAR%3Binac%3D0%3Bindex_postcard%3D0%3Blabel_click%3Dundef%3Bno_rooms%3D1%3Boffset%3D0%3Bpostcard%3D0%3Braw_dest_type%3Dcity%3Broom1%3DA%252CA%3Bsb_price_type%3Dtotal%3Bsearch_selected%3D1%3Bshw_aparth%3D1%3Bslp_r_match%3D0%3Bsrc%3Dsearchresults%3Bsrc_elem%3Dsb%3Bsrpvid%3Dd4b27cd4b6490100%3Bss%3DParis%252C%2520%25C3%258Ele-de-France%252C%2520France%3Bss_all%3D0%3Bss_raw%3Dparis%3Bssb%3Dempty%3Bsshis%3D0%3Bssne%3DRouen%3Bssne_untouched%3DRouen%3Btop_ufis%3D1%26%3B&ss=Amiens%2C+Picardie%2C+France&is_ski_area=&ssne=Paris&ssne_untouched=Paris&city=-1456928&checkin_year=&checkin_month=&checkout_year=&checkout_month=&group_adults=2&group_children=0&no_rooms=1&from_sf=1&ss_raw=amiens&ac_position=0&ac_langcode=fr&ac_click_type=b&dest_id=-1407447&dest_type=city&place_id_lat=49.895454&place_id_lon=2.299816&search_pageview_id=d4b27cd4b6490100&search_selected=true&search_pageview_id=d4b27cd4b6490100&ac_suggestion_list_length=5&ac_suggestion_theme_list_length=0&nflt=ht_id%3D204"
url_lille = "https://www.booking.com/searchresults.fr.html?label=gen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB&sid=a33dd63e1835e1ec9af64c715c2bc273&aid=304142&src=searchresults&error_url=https%3A%2F%2Fwww.booking.com%2Fsearchresults.fr.html%3Faid%3D304142%3Blabel%3Dgen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB%3Bsid%3Da33dd63e1835e1ec9af64c715c2bc273%3Btmpl%3Dsearchresults%3Bac_click_type%3Db%3Bac_position%3D0%3Bcity%3D-1456928%3Bclass_interval%3D1%3Bdest_id%3D-1407447%3Bdest_type%3Dcity%3Bdtdisc%3D0%3Bfrom_sf%3D1%3Bgroup_adults%3D2%3Bgroup_children%3D0%3Binac%3D0%3Bindex_postcard%3D0%3Blabel_click%3Dundef%3Bno_rooms%3D1%3Boffset%3D0%3Bpostcard%3D0%3Braw_dest_type%3Dcity%3Broom1%3DA%252CA%3Bsb_price_type%3Dtotal%3Bsearch_selected%3D1%3Bshw_aparth%3D1%3Bslp_r_match%3D0%3Bsrc%3Dsearchresults%3Bsrc_elem%3Dsb%3Bsrpvid%3D51b37cfcc6f2016c%3Bss%3DAmiens%252C%2520Picardie%252C%2520France%3Bss_all%3D0%3Bss_raw%3Damiens%3Bssb%3Dempty%3Bsshis%3D0%3Bssne%3DParis%3Bssne_untouched%3DParis%3Btop_ufis%3D1%26%3B&ss=Lille%2C+Nord-Pas-de-Calais%2C+France&is_ski_area=0&ssne=Amiens&ssne_untouched=Amiens&city=-1407447&checkin_year=&checkin_month=&checkout_year=&checkout_month=&group_adults=2&group_children=0&no_rooms=1&from_sf=1&ss_raw=lille&ac_position=0&ac_langcode=fr&ac_click_type=b&dest_id=-1447079&dest_type=city&iata=LIL&place_id_lat=50.636845&place_id_lon=3.063454&search_pageview_id=51b37cfcc6f2016c&search_selected=true&nflt=ht_id%3D204"
url_strasbourg = "https://www.booking.com/searchresults.fr.html?label=gen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB&sid=430afe453916ce7d5412983b211cdb2d&aid=304142&src=searchresults&error_url=https%3A%2F%2Fwww.booking.com%2Fsearchresults.fr.html%3Faid%3D304142%3Blabel%3Dgen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB%3Bsid%3D430afe453916ce7d5412983b211cdb2d%3Btmpl%3Dsearchresults%3Bac_click_type%3Db%3Bac_position%3D0%3Bcity%3D-1407447%3Bclass_interval%3D1%3Bdest_id%3D-1447079%3Bdest_type%3Dcity%3Bdtdisc%3D0%3Bfrom_sf%3D1%3Bgroup_adults%3D2%3Bgroup_children%3D0%3Biata%3DLIL%3Binac%3D0%3Bindex_postcard%3D0%3Blabel_click%3Dundef%3Bnflt%3Dht_id%253D204%3Bno_rooms%3D1%3Boffset%3D0%3Bpostcard%3D0%3Braw_dest_type%3Dcity%3Broom1%3DA%252CA%3Bsb_price_type%3Dtotal%3Bsearch_selected%3D1%3Bshw_aparth%3D1%3Bslp_r_match%3D0%3Bsrc%3Dsearchresults%3Bsrpvid%3D6a3a2fbd4d1d00b2%3Bss%3DLille%252C%2520Nord-Pas-de-Calais%252C%2520France%3Bss_all%3D0%3Bss_raw%3Dlille%3Bssb%3Dempty%3Bsshis%3D0%3Bssne%3DAmiens%3Bssne_untouched%3DAmiens%3Btop_ufis%3D1%26%3B&ss=Strasbourg%2C+Alsace%2C+France&is_ski_area=&ssne=Lille&ssne_untouched=Lille&city=-1447079&checkin_year=&checkin_month=&checkout_year=&checkout_month=&group_adults=2&group_children=0&no_rooms=1&from_sf=1&ss_raw=stras&ac_position=0&ac_langcode=fr&ac_click_type=b&dest_id=-1471697&dest_type=city&iata=SXB&place_id_lat=48.5834&place_id_lon=7.7458&search_pageview_id=6a3a2fbd4d1d00b2&search_selected=true&search_pageview_id=6a3a2fbd4d1d00b2&ac_suggestion_list_length=5&ac_suggestion_theme_list_length=0&nflt=ht_id%3D204"
url_chateau_du_haut_koenigsbourg = "https://www.booking.com/searchresults.fr.html?label=gen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB&sid=430afe453916ce7d5412983b211cdb2d&aid=304142&src=searchresults&error_url=https%3A%2F%2Fwww.booking.com%2Fsearchresults.fr.html%3Faid%3D304142%3Blabel%3Dgen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB%3Bsid%3D430afe453916ce7d5412983b211cdb2d%3Btmpl%3Dsearchresults%3Bac_click_type%3Db%3Bac_position%3D0%3Bcity%3D-1447079%3Bclass_interval%3D1%3Bdest_id%3D-1471697%3Bdest_type%3Dcity%3Bdtdisc%3D0%3Bfrom_sf%3D1%3Bgroup_adults%3D2%3Bgroup_children%3D0%3Biata%3DSXB%3Binac%3D0%3Bindex_postcard%3D0%3Blabel_click%3Dundef%3Bno_rooms%3D1%3Boffset%3D0%3Bpostcard%3D0%3Braw_dest_type%3Dcity%3Broom1%3DA%252CA%3Bsb_price_type%3Dtotal%3Bsearch_selected%3D1%3Bshw_aparth%3D1%3Bslp_r_match%3D0%3Bsrc%3Dsearchresults%3Bsrc_elem%3Dsb%3Bsrpvid%3D03962fc6100f0037%3Bss%3DStrasbourg%252C%2520Alsace%252C%2520France%3Bss_all%3D0%3Bss_raw%3Dstras%3Bssb%3Dempty%3Bsshis%3D0%3Bssne%3DLille%3Bssne_untouched%3DLille%3Btop_ufis%3D1%26%3B&ss=Ch%C3%A2teau+du+Haut-K%C5%93nigsbourg%2C+Saint-Hippolyte%2C+Alsace%2C+France&is_ski_area=&ssne=Strasbourg&ssne_untouched=Strasbourg&city=-1471697&checkin_year=&checkin_month=&checkout_year=&checkout_month=&group_adults=2&group_children=0&no_rooms=1&from_sf=1&ss_raw=chateau+koe&ac_position=2&ac_langcode=fr&ac_click_type=b&dest_id=204055&dest_type=landmark&place_id_lat=48.24931&place_id_lon=7.34445&search_pageview_id=03962fc6100f0037&search_selected=true&search_pageview_id=03962fc6100f0037&ac_suggestion_list_length=5&ac_suggestion_theme_list_length=0&nflt=ht_id%3D204"
url_colmar = "https://www.booking.com/searchresults.fr.html?label=gen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB&sid=430afe453916ce7d5412983b211cdb2d&aid=304142&src=searchresults&error_url=https%3A%2F%2Fwww.booking.com%2Fsearchresults.fr.html%3Faid%3D304142%3Blabel%3Dgen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB%3Bsid%3D430afe453916ce7d5412983b211cdb2d%3Btmpl%3Dsearchresults%3Bac_click_type%3Db%3Bac_position%3D2%3Bcity%3D-1471697%3Bclass_interval%3D1%3Bdest_id%3D204055%3Bdest_type%3Dlandmark%3Bdtdisc%3D0%3Bfrom_sf%3D1%3Bgroup_adults%3D2%3Bgroup_children%3D0%3Binac%3D0%3Bindex_postcard%3D0%3Blabel_click%3Dundef%3Bno_rooms%3D1%3Boffset%3D0%3Bpostcard%3D0%3Braw_dest_type%3Dlandmark%3Broom1%3DA%252CA%3Bsb_price_type%3Dtotal%3Bsearch_selected%3D1%3Bshw_aparth%3D1%3Bslp_r_match%3D0%3Bsrc%3Dsearchresults%3Bsrc_elem%3Dsb%3Bsrpvid%3D419b2fe26add0110%3Bss%3DCh%25C3%25A2teau%2520du%2520Haut-K%25C5%2593nigsbourg%252C%2520Saint-Hippolyte%252C%2520Alsace%252C%2520France%3Bss_all%3D0%3Bss_raw%3Dchateau%2520koe%3Bssb%3Dempty%3Bsshis%3D0%3Bssne%3DStrasbourg%3Bssne_untouched%3DStrasbourg%3Btop_ufis%3D1%26%3B&ss=Colmar%2C+Alsace%2C+France&is_ski_area=&ssne=Ch%C3%A2teau+du+Haut-K%C5%93nigsbourg&ssne_untouched=Ch%C3%A2teau+du+Haut-K%C5%93nigsbourg&checkin_year=&checkin_month=&checkout_year=&checkout_month=&group_adults=2&group_children=0&no_rooms=1&from_sf=1&ss_raw=colma&ac_position=0&ac_langcode=fr&ac_click_type=b&dest_id=-1421049&dest_type=city&place_id_lat=48.079155&place_id_lon=7.357718&search_pageview_id=419b2fe26add0110&search_selected=true&search_pageview_id=419b2fe26add0110&ac_suggestion_list_length=5&ac_suggestion_theme_list_length=0&nflt=ht_id%3D204"
url_eguisheim = "https://www.booking.com/searchresults.fr.html?label=gen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB&sid=430afe453916ce7d5412983b211cdb2d&aid=304142&src=searchresults&error_url=https%3A%2F%2Fwww.booking.com%2Fsearchresults.fr.html%3Faid%3D304142%3Blabel%3Dgen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB%3Bsid%3D430afe453916ce7d5412983b211cdb2d%3Btmpl%3Dsearchresults%3Bac_click_type%3Db%3Bac_position%3D0%3Bclass_interval%3D1%3Bdest_id%3D-1421049%3Bdest_type%3Dcity%3Bdtdisc%3D0%3Bfrom_sf%3D1%3Bgroup_adults%3D2%3Bgroup_children%3D0%3Binac%3D0%3Bindex_postcard%3D0%3Blabel_click%3Dundef%3Bno_rooms%3D1%3Boffset%3D0%3Bpostcard%3D0%3Braw_dest_type%3Dcity%3Broom1%3DA%252CA%3Bsb_price_type%3Dtotal%3Bsearch_selected%3D1%3Bshw_aparth%3D1%3Bslp_r_match%3D0%3Bsrc%3Dsearchresults%3Bsrc_elem%3Dsb%3Bsrpvid%3Dae792feada870020%3Bss%3DColmar%252C%2520Alsace%252C%2520France%3Bss_all%3D0%3Bss_raw%3Dcolma%3Bssb%3Dempty%3Bsshis%3D0%3Bssne%3DCh%25C3%25A2teau%2520du%2520Haut-K%25C5%2593nigsbourg%3Bssne_untouched%3DCh%25C3%25A2teau%2520du%2520Haut-K%25C5%2593nigsbourg%3Btop_ufis%3D1%26%3B&ss=Eguisheim%2C+Alsace%2C+France&is_ski_area=&ssne=Colmar&ssne_untouched=Colmar&city=-1421049&checkin_year=&checkin_month=&checkout_year=&checkout_month=&group_adults=2&group_children=0&no_rooms=1&from_sf=1&ss_raw=eguisheim&ac_position=0&ac_langcode=fr&ac_click_type=b&dest_id=-1425030&dest_type=city&place_id_lat=48.04231&place_id_lon=7.30543&search_pageview_id=ae792feada870020&search_selected=true&search_pageview_id=ae792feada870020&ac_suggestion_list_length=5&ac_suggestion_theme_list_length=0&nflt=ht_id%3D204"
url_besançon = "https://www.booking.com/searchresults.fr.html?label=gen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB&sid=430afe453916ce7d5412983b211cdb2d&aid=304142&src=searchresults&error_url=https%3A%2F%2Fwww.booking.com%2Fsearchresults.fr.html%3Faid%3D304142%3Blabel%3Dgen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB%3Bsid%3D430afe453916ce7d5412983b211cdb2d%3Btmpl%3Dsearchresults%3Bac_click_type%3Db%3Bac_position%3D0%3Bcity%3D-1421049%3Bclass_interval%3D1%3Bdest_id%3D-1425030%3Bdest_type%3Dcity%3Bdtdisc%3D0%3Bfrom_sf%3D1%3Bgroup_adults%3D2%3Bgroup_children%3D0%3Binac%3D0%3Bindex_postcard%3D0%3Blabel_click%3Dundef%3Bno_rooms%3D1%3Boffset%3D0%3Bpostcard%3D0%3Braw_dest_type%3Dcity%3Broom1%3DA%252CA%3Bsb_price_type%3Dtotal%3Bsearch_selected%3D1%3Bshw_aparth%3D1%3Bslp_r_match%3D0%3Bsrc%3Dsearchresults%3Bsrc_elem%3Dsb%3Bsrpvid%3D38d92ff4808600b8%3Bss%3DEguisheim%252C%2520Alsace%252C%2520France%3Bss_all%3D0%3Bss_raw%3Deguisheim%3Bssb%3Dempty%3Bsshis%3D0%3Bssne%3DColmar%3Bssne_untouched%3DColmar%3Btop_ufis%3D1%26%3B&ss=Besan%C3%A7on%2C+France&is_ski_area=&ssne=Eguisheim&ssne_untouched=Eguisheim&city=-1425030&checkin_year=&checkin_month=&checkout_year=&checkout_month=&group_adults=2&group_children=0&no_rooms=1&from_sf=1&ss_raw=besan_on&ac_position=1&ac_click_type=g&dest_id=ChIJCxPu8PpijUcRMO0TszTOCQQ&dest_type=city&place_id=ChIJCxPu8PpijUcRMO0TszTOCQQ&place_id_lat=47.237829&place_id_lon=6.0240539&place_types=locality%2Cpolitical%2Cgeocode&search_pageview_id=38d92ff4808600b8&search_selected=true&nflt=ht_id%3D204"
url_dijon = "https://www.booking.com/searchresults.fr.html?label=gen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB&sid=430afe453916ce7d5412983b211cdb2d&aid=304142&src=searchresults&error_url=https%3A%2F%2Fwww.booking.com%2Fsearchresults.fr.html%3Faid%3D304142%3Blabel%3Dgen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB%3Bsid%3D430afe453916ce7d5412983b211cdb2d%3Btmpl%3Dsearchresults%3Bac_click_type%3Dg%3Bclass_interval%3D1%3Bdtdisc%3D0%3Bfrom_sf%3D1%3Bgroup_adults%3D2%3Bgroup_children%3D0%3Binac%3D0%3Bindex_postcard%3D0%3Blabel_click%3Dundef%3Bno_rooms%3D1%3Boffset%3D0%3Bplace_id%3DChIJCxPu8PpijUcRMO0TszTOCQQ%3Bplace_id_lat%3D47.237829%3Bplace_id_lon%3D6.0240539%3Bpostcard%3D0%3Braw_dest_type%3Dcity%3Broom1%3DA%252CA%3Bsb_price_type%3Dtotal%3Bsearch_selected%3D1%3Bshw_aparth%3D1%3Bslp_r_match%3D0%3Bsrc%3Dsearchresults%3Bsrc_elem%3Dsb%3Bsrpvid%3D00a33006b655013c%3Bss%3DBesan%25C3%25A7on%252C%2520France%3Bss_all%3D0%3Bss_raw%3Dbesan_on%3Bssb%3Dempty%3Bsshis%3D0%3Bssne%3DEguisheim%3Bssne_untouched%3DEguisheim%3Btop_ufis%3D1%26%3B&ss=Dijon%2C+Bourgogne%2C+France&is_ski_area=&ssne=Besan%C3%A7on%2C+France&ssne_untouched=Besan%C3%A7on%2C+France&checkin_year=&checkin_month=&checkout_year=&checkout_month=&group_adults=2&group_children=0&no_rooms=1&from_sf=1&ss_raw=dijon&ac_position=0&ac_langcode=fr&ac_click_type=b&dest_id=-1423981&dest_type=city&place_id_lat=47.321762&place_id_lon=5.039809&search_pageview_id=00a33006b655013c&search_selected=true&search_pageview_id=00a33006b655013c&ac_suggestion_list_length=5&ac_suggestion_theme_list_length=0&nflt=ht_id%3D204"
url_annecy = "https://www.booking.com/searchresults.fr.html?label=gen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB&sid=430afe453916ce7d5412983b211cdb2d&aid=304142&src=searchresults&error_url=https%3A%2F%2Fwww.booking.com%2Fsearchresults.fr.html%3Faid%3D304142%3Blabel%3Dgen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB%3Bsid%3D430afe453916ce7d5412983b211cdb2d%3Btmpl%3Dsearchresults%3Bac_click_type%3Db%3Bac_position%3D0%3Bclass_interval%3D1%3Bdest_id%3D-1423981%3Bdest_type%3Dcity%3Bdtdisc%3D0%3Bfrom_sf%3D1%3Bgroup_adults%3D2%3Bgroup_children%3D0%3Binac%3D0%3Bindex_postcard%3D0%3Blabel_click%3Dundef%3Bnflt%3Dht_id%253D204%3Bno_rooms%3D1%3Boffset%3D0%3Bpostcard%3D0%3Braw_dest_type%3Dcity%3Broom1%3DA%252CA%3Bsb_price_type%3Dtotal%3Bsearch_selected%3D1%3Bshw_aparth%3D1%3Bslp_r_match%3D0%3Bsrc%3Dsearchresults%3Bsrpvid%3Dedf7301afb1b00fc%3Bss%3DDijon%252C%2520Bourgogne%252C%2520France%3Bss_all%3D0%3Bss_raw%3Ddijon%3Bssb%3Dempty%3Bsshis%3D0%3Bssne%3DBesan%25C3%25A7on%252C%2520France%3Bssne_untouched%3DBesan%25C3%25A7on%252C%2520France%3Btop_ufis%3D1%26%3B&ss=Annecy%2C+Rh%C3%B4ne-Alpes%2C+France&is_ski_area=&ssne=Dijon&ssne_untouched=Dijon&city=-1423981&checkin_year=&checkin_month=&checkout_year=&checkout_month=&group_adults=2&group_children=0&no_rooms=1&from_sf=1&ss_raw=annecy&ac_position=0&ac_langcode=fr&ac_click_type=b&dest_id=-1407760&dest_type=city&iata=NCY&place_id_lat=45.899727&place_id_lon=6.127196&search_pageview_id=edf7301afb1b00fc&search_selected=true&search_pageview_id=edf7301afb1b00fc&ac_suggestion_list_length=5&ac_suggestion_theme_list_length=0&nflt=ht_id%3D204"
url_grenoble = "https://www.booking.com/searchresults.fr.html?label=gen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB&sid=430afe453916ce7d5412983b211cdb2d&aid=304142&src=searchresults&error_url=https%3A%2F%2Fwww.booking.com%2Fsearchresults.fr.html%3Faid%3D304142%3Blabel%3Dgen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB%3Bsid%3D430afe453916ce7d5412983b211cdb2d%3Btmpl%3Dsearchresults%3Bac_click_type%3Db%3Bac_position%3D0%3Bcity%3D-1423981%3Bclass_interval%3D1%3Bdest_id%3D-1407760%3Bdest_type%3Dcity%3Bdtdisc%3D0%3Bfrom_sf%3D1%3Bgroup_adults%3D2%3Bgroup_children%3D0%3Biata%3DNCY%3Binac%3D0%3Bindex_postcard%3D0%3Blabel_click%3Dundef%3Bno_rooms%3D1%3Boffset%3D0%3Bpostcard%3D0%3Braw_dest_type%3Dcity%3Broom1%3DA%252CA%3Bsb_price_type%3Dtotal%3Bsearch_selected%3D1%3Bshw_aparth%3D1%3Bslp_r_match%3D0%3Bsrc%3Dsearchresults%3Bsrc_elem%3Dsb%3Bsrpvid%3D11c23023f64a00d8%3Bss%3DAnnecy%252C%2520Rh%25C3%25B4ne-Alpes%252C%2520France%3Bss_all%3D0%3Bss_raw%3Dannecy%3Bssb%3Dempty%3Bsshis%3D0%3Bssne%3DDijon%3Bssne_untouched%3DDijon%3Btop_ufis%3D1%26%3B&ss=Grenoble%2C+Rh%C3%B4ne-Alpes%2C+France&is_ski_area=&ssne=Annecy&ssne_untouched=Annecy&city=-1407760&checkin_year=&checkin_month=&checkout_year=&checkout_month=&group_adults=2&group_children=0&no_rooms=1&from_sf=1&ss_raw=grenoble&ac_position=0&ac_langcode=fr&ac_click_type=b&dest_id=-1430647&dest_type=city&place_id_lat=45.192505&place_id_lon=5.724973&search_pageview_id=11c23023f64a00d8&search_selected=true&search_pageview_id=11c23023f64a00d8&ac_suggestion_list_length=5&ac_suggestion_theme_list_length=0&nflt=ht_id%3D204"
url_lyon = "https://www.booking.com/searchresults.fr.html?label=gen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB&sid=430afe453916ce7d5412983b211cdb2d&aid=304142&src=searchresults&error_url=https%3A%2F%2Fwww.booking.com%2Fsearchresults.fr.html%3Faid%3D304142%3Blabel%3Dgen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB%3Bsid%3D430afe453916ce7d5412983b211cdb2d%3Btmpl%3Dsearchresults%3Bac_click_type%3Db%3Bac_position%3D0%3Bcity%3D-1407760%3Bclass_interval%3D1%3Bdest_id%3D-1430647%3Bdest_type%3Dcity%3Bdtdisc%3D0%3Bfrom_sf%3D1%3Bgroup_adults%3D2%3Bgroup_children%3D0%3Binac%3D0%3Bindex_postcard%3D0%3Blabel_click%3Dundef%3Bno_rooms%3D1%3Boffset%3D0%3Bpostcard%3D0%3Braw_dest_type%3Dcity%3Broom1%3DA%252CA%3Bsb_price_type%3Dtotal%3Bsearch_selected%3D1%3Bshw_aparth%3D1%3Bslp_r_match%3D0%3Bsrc%3Dsearchresults%3Bsrc_elem%3Dsb%3Bsrpvid%3D2a54302c517d00c5%3Bss%3DGrenoble%252C%2520Rh%25C3%25B4ne-Alpes%252C%2520France%3Bss_all%3D0%3Bss_raw%3Dgrenoble%3Bssb%3Dempty%3Bsshis%3D0%3Bssne%3DAnnecy%3Bssne_untouched%3DAnnecy%3Btop_ufis%3D1%26%3B&ss=Lyon%2C+Rh%C3%B4ne-Alpes%2C+France&is_ski_area=0&ssne=Grenoble&ssne_untouched=Grenoble&city=-1430647&checkin_year=&checkin_month=&checkout_year=&checkout_month=&group_adults=2&group_children=0&no_rooms=1&from_sf=1&ss_raw=lyon&ac_position=0&ac_langcode=fr&ac_click_type=b&dest_id=-1448468&dest_type=city&iata=LYS&place_id_lat=45.757767&place_id_lon=4.832204&search_pageview_id=2a54302c517d00c5&search_selected=true&nflt=ht_id%3D204"
url_gorges_du_verdon = "https://www.booking.com/searchresults.fr.html?label=gen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB&sid=430afe453916ce7d5412983b211cdb2d&aid=304142&src=searchresults&error_url=https%3A%2F%2Fwww.booking.com%2Fsearchresults.fr.html%3Faid%3D304142%3Blabel%3Dgen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB%3Bsid%3D430afe453916ce7d5412983b211cdb2d%3Btmpl%3Dsearchresults%3Bac_click_type%3Db%3Bac_position%3D0%3Bcity%3D-1430647%3Bclass_interval%3D1%3Bdest_id%3D-1448468%3Bdest_type%3Dcity%3Bdtdisc%3D0%3Bfrom_sf%3D1%3Bgroup_adults%3D2%3Bgroup_children%3D0%3Biata%3DLYS%3Binac%3D0%3Bindex_postcard%3D0%3Blabel_click%3Dundef%3Bno_rooms%3D1%3Boffset%3D0%3Bpostcard%3D0%3Braw_dest_type%3Dcity%3Broom1%3DA%252CA%3Bsb_price_type%3Dtotal%3Bsearch_selected%3D1%3Bshw_aparth%3D1%3Bslp_r_match%3D0%3Bsrc%3Dsearchresults%3Bsrc_elem%3Dsb%3Bsrpvid%3D52ad305ee29d00fc%3Bss%3DLyon%252C%2520Rh%25C3%25B4ne-Alpes%252C%2520France%3Bss_all%3D0%3Bss_raw%3Dlyon%3Bssb%3Dempty%3Bsshis%3D0%3Bssne%3DGrenoble%3Bssne_untouched%3DGrenoble%3Btop_ufis%3D1%26%3B&ss=Gorges+du+Verdon%2C+France&is_ski_area=&ssne=Lyon&ssne_untouched=Lyon&city=-1448468&checkin_year=&checkin_month=&checkout_year=&checkout_month=&group_adults=2&group_children=0&no_rooms=1&from_sf=1&ss_raw=gorges+verdon&ac_position=0&ac_langcode=fr&ac_click_type=b&dest_id=2746&dest_type=region&place_id_lat=43.78456&place_id_lon=6.126273&search_pageview_id=52ad305ee29d00fc&search_selected=true&search_pageview_id=52ad305ee29d00fc&ac_suggestion_list_length=5&ac_suggestion_theme_list_length=0&nflt=ht_id%3D204"
url_bormes_les_mimosas = "https://www.booking.com/searchresults.fr.html?label=gen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB&sid=430afe453916ce7d5412983b211cdb2d&aid=304142&src=searchresults&error_url=https%3A%2F%2Fwww.booking.com%2Fsearchresults.fr.html%3Faid%3D304142%3Blabel%3Dgen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB%3Bsid%3D430afe453916ce7d5412983b211cdb2d%3Btmpl%3Dsearchresults%3Bac_click_type%3Db%3Bac_position%3D0%3Bcity%3D-1448468%3Bclass_interval%3D1%3Bdest_id%3D2746%3Bdest_type%3Dregion%3Bdtdisc%3D0%3Bfrom_sf%3D1%3Bgroup_adults%3D2%3Bgroup_children%3D0%3Binac%3D0%3Bindex_postcard%3D0%3Blabel_click%3Dundef%3Bno_rooms%3D1%3Boffset%3D0%3Bpostcard%3D0%3Braw_dest_type%3Dregion%3Broom1%3DA%252CA%3Bsb_price_type%3Dtotal%3Bsearch_selected%3D1%3Bshw_aparth%3D1%3Bslp_r_match%3D0%3Bsrc%3Dsearchresults%3Bsrc_elem%3Dsb%3Bsrpvid%3D48953068780401a9%3Bss%3DGorges%2520du%2520Verdon%252C%2520France%3Bss_all%3D0%3Bss_raw%3Dgorges%2520verdon%3Bssb%3Dempty%3Bsshis%3D0%3Bssne%3DLyon%3Bssne_untouched%3DLyon%3Btop_ufis%3D1%26%3B&ss=Bormes-les-Mimosas%2C+Provence-Alpes-C%C3%B4te+d%27Azur%2C+France&is_ski_area=&ssne=Gorges+du+Verdon&ssne_untouched=Gorges+du+Verdon&checkin_year=&checkin_month=&checkout_year=&checkout_month=&group_adults=2&group_children=0&no_rooms=1&from_sf=1&ss_raw=bormes+les+mim&ac_position=0&ac_langcode=fr&ac_click_type=b&dest_id=-1413801&dest_type=city&place_id_lat=43.1507&place_id_lon=6.34193&search_pageview_id=48953068780401a9&search_selected=true&search_pageview_id=48953068780401a9&ac_suggestion_list_length=5&ac_suggestion_theme_list_length=0&nflt=ht_id%3D204"
url_cassis = "https://www.booking.com/searchresults.fr.html?label=gen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB&sid=430afe453916ce7d5412983b211cdb2d&aid=304142&src=searchresults&error_url=https%3A%2F%2Fwww.booking.com%2Fsearchresults.fr.html%3Faid%3D304142%3Blabel%3Dgen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB%3Bsid%3D430afe453916ce7d5412983b211cdb2d%3Btmpl%3Dsearchresults%3Bac_click_type%3Db%3Bac_position%3D0%3Bclass_interval%3D1%3Bdest_id%3D-1413801%3Bdest_type%3Dcity%3Bdtdisc%3D0%3Bfrom_sf%3D1%3Bgroup_adults%3D2%3Bgroup_children%3D0%3Binac%3D0%3Bindex_postcard%3D0%3Blabel_click%3Dundef%3Bno_rooms%3D1%3Boffset%3D0%3Bpostcard%3D0%3Braw_dest_type%3Dcity%3Broom1%3DA%252CA%3Bsb_price_type%3Dtotal%3Bsearch_selected%3D1%3Bshw_aparth%3D1%3Bslp_r_match%3D0%3Bsrc%3Dsearchresults%3Bsrc_elem%3Dsb%3Bsrpvid%3D465d3073667a0156%3Bss%3DBormes-les-Mimosas%252C%2520Provence-Alpes-C%25C3%25B4te%2520d%2527Azur%252C%2520France%3Bss_all%3D0%3Bss_raw%3Dbormes%2520les%2520mim%3Bssb%3Dempty%3Bsshis%3D0%3Bssne%3DGorges%2520du%2520Verdon%3Bssne_untouched%3DGorges%2520du%2520Verdon%3Btop_ufis%3D1%26%3B&ss=Cassis%2C+Provence-Alpes-C%C3%B4te+d%27Azur%2C+France&is_ski_area=0&ssne=Bormes-les-Mimosas&ssne_untouched=Bormes-les-Mimosas&city=-1413801&checkin_year=&checkin_month=&checkout_year=&checkout_month=&group_adults=2&group_children=0&no_rooms=1&from_sf=1&ss_raw=cassis&ac_position=0&ac_langcode=fr&ac_click_type=b&dest_id=-1416912&dest_type=city&place_id_lat=43.21513&place_id_lon=5.53712&search_pageview_id=465d3073667a0156&search_selected=true&nflt=ht_id%3D204"
url_marseille = "https://www.booking.com/searchresults.fr.html?label=gen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB&sid=430afe453916ce7d5412983b211cdb2d&aid=304142&src=searchresults&error_url=https%3A%2F%2Fwww.booking.com%2Fsearchresults.fr.html%3Faid%3D304142%3Blabel%3Dgen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB%3Bsid%3D430afe453916ce7d5412983b211cdb2d%3Btmpl%3Dsearchresults%3Bac_click_type%3Db%3Bac_position%3D0%3Bcity%3D-1413801%3Bclass_interval%3D1%3Bdest_id%3D-1416912%3Bdest_type%3Dcity%3Bdtdisc%3D0%3Bfrom_sf%3D1%3Bgroup_adults%3D2%3Bgroup_children%3D0%3Binac%3D0%3Bindex_postcard%3D0%3Blabel_click%3Dundef%3Bno_rooms%3D1%3Boffset%3D0%3Bpostcard%3D0%3Braw_dest_type%3Dcity%3Broom1%3DA%252CA%3Bsb_price_type%3Dtotal%3Bsearch_selected%3D1%3Bshw_aparth%3D1%3Bslp_r_match%3D0%3Bsrc%3Dsearchresults%3Bsrc_elem%3Dsb%3Bsrpvid%3Ddd9e307ab428001b%3Bss%3DCassis%252C%2520Provence-Alpes-C%25C3%25B4te%2520d%2527Azur%252C%2520France%3Bss_all%3D0%3Bss_raw%3Dcassis%3Bssb%3Dempty%3Bsshis%3D0%3Bssne%3DBormes-les-Mimosas%3Bssne_untouched%3DBormes-les-Mimosas%3Btop_ufis%3D1%26%3B&ss=marseille&is_ski_area=0&ssne=Cassis&ssne_untouched=Cassis&city=-1416912&checkin_year=&checkin_month=&checkout_year=&checkout_month=&group_adults=2&group_children=0&no_rooms=1&from_sf=1&nflt=ht_id%3D204"
url_aix_en_provence = "https://www.booking.com/searchresults.fr.html?label=gen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB&sid=430afe453916ce7d5412983b211cdb2d&aid=304142&src=searchresults&error_url=https%3A%2F%2Fwww.booking.com%2Fsearchresults.fr.html%3Faid%3D304142%3Blabel%3Dgen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB%3Bsid%3D430afe453916ce7d5412983b211cdb2d%3Btmpl%3Dsearchresults%3Bcity%3D-1416912%3Bclass_interval%3D1%3Bdest_id%3D-1416912%3Bdest_type%3Dcity%3Bdtdisc%3D0%3Bfrom_sf%3D1%3Bgroup_adults%3D2%3Bgroup_children%3D0%3Binac%3D0%3Bindex_postcard%3D0%3Blabel_click%3Dundef%3Bno_rooms%3D1%3Boffset%3D0%3Bpostcard%3D0%3Broom1%3DA%252CA%3Bsb_price_type%3Dtotal%3Bshw_aparth%3D1%3Bslp_r_match%3D0%3Bsrc%3Dsearchresults%3Bsrc_elem%3Dsb%3Bsrpvid%3D5740308111d1001d%3Bss%3Dmarseille%3Bss_all%3D0%3Bssb%3Dempty%3Bsshis%3D0%3Bssne%3DCassis%3Bssne_untouched%3DCassis%3Btop_ufis%3D1%26%3B&ss=Aix-en-Provence%2C+Provence-Alpes-C%C3%B4te+d%27Azur%2C+France&is_ski_area=&ssne=Cassis&ssne_untouched=Cassis&city=-1416912&checkin_year=&checkin_month=&checkout_year=&checkout_month=&group_adults=2&group_children=0&no_rooms=1&from_sf=1&ss_raw=aix+en+provence&ac_position=0&ac_langcode=fr&ac_click_type=b&dest_id=-1406939&dest_type=city&place_id_lat=43.526314&place_id_lon=5.445464&search_pageview_id=5740308111d1001d&search_selected=true&search_pageview_id=5740308111d1001d&ac_suggestion_list_length=5&ac_suggestion_theme_list_length=0&nflt=ht_id%3D204"
url_avignon = "https://www.booking.com/searchresults.fr.html?label=gen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB&sid=430afe453916ce7d5412983b211cdb2d&aid=304142&src=searchresults&error_url=https%3A%2F%2Fwww.booking.com%2Fsearchresults.fr.html%3Faid%3D304142%3Blabel%3Dgen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB%3Bsid%3D430afe453916ce7d5412983b211cdb2d%3Btmpl%3Dsearchresults%3Bac_click_type%3Db%3Bac_position%3D0%3Bcity%3D-1416912%3Bclass_interval%3D1%3Bdest_id%3D-1406939%3Bdest_type%3Dcity%3Bdtdisc%3D0%3Bfrom_sf%3D1%3Bgroup_adults%3D2%3Bgroup_children%3D0%3Binac%3D0%3Bindex_postcard%3D0%3Blabel_click%3Dundef%3Bno_rooms%3D1%3Boffset%3D0%3Bpostcard%3D0%3Braw_dest_type%3Dcity%3Broom1%3DA%252CA%3Bsb_price_type%3Dtotal%3Bsearch_selected%3D1%3Bshw_aparth%3D1%3Bslp_r_match%3D0%3Bsrc%3Dsearchresults%3Bsrc_elem%3Dsb%3Bsrpvid%3D592e3089d8650155%3Bss%3DAix-en-Provence%252C%2520Provence-Alpes-C%25C3%25B4te%2520d%2527Azur%252C%2520France%3Bss_all%3D0%3Bss_raw%3Daix%2520en%2520provence%3Bssb%3Dempty%3Bsshis%3D0%3Bssne%3DCassis%3Bssne_untouched%3DCassis%3Btop_ufis%3D1%26%3B&ss=Avignon%2C+Provence-Alpes-C%C3%B4te+d%27Azur%2C+France&is_ski_area=&ssne=Aix-en-Provence&ssne_untouched=Aix-en-Provence&city=-1406939&checkin_year=&checkin_month=&checkout_year=&checkout_month=&group_adults=2&group_children=0&no_rooms=1&from_sf=1&ss_raw=avignon&ac_position=0&ac_langcode=fr&ac_click_type=b&dest_id=-1409631&dest_type=city&iata=AVN&place_id_lat=43.949&place_id_lon=4.807508&search_pageview_id=592e3089d8650155&search_selected=true&search_pageview_id=592e3089d8650155&ac_suggestion_list_length=5&ac_suggestion_theme_list_length=0&nflt=ht_id%3D204"
url_uzes = "https://www.booking.com/searchresults.fr.html?label=gen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB&sid=430afe453916ce7d5412983b211cdb2d&aid=304142&src=searchresults&error_url=https%3A%2F%2Fwww.booking.com%2Fsearchresults.fr.html%3Faid%3D304142%3Blabel%3Dgen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB%3Bsid%3D430afe453916ce7d5412983b211cdb2d%3Btmpl%3Dsearchresults%3Bac_click_type%3Db%3Bac_position%3D0%3Bcity%3D-1406939%3Bclass_interval%3D1%3Bdest_id%3D-1409631%3Bdest_type%3Dcity%3Bdtdisc%3D0%3Bfrom_sf%3D1%3Bgroup_adults%3D2%3Bgroup_children%3D0%3Biata%3DAVN%3Binac%3D0%3Bindex_postcard%3D0%3Blabel_click%3Dundef%3Bno_rooms%3D1%3Boffset%3D0%3Bpostcard%3D0%3Braw_dest_type%3Dcity%3Broom1%3DA%252CA%3Bsb_price_type%3Dtotal%3Bsearch_selected%3D1%3Bshw_aparth%3D1%3Bslp_r_match%3D0%3Bsrc%3Dsearchresults%3Bsrc_elem%3Dsb%3Bsrpvid%3Dbd4030933e150131%3Bss%3DAvignon%252C%2520Provence-Alpes-C%25C3%25B4te%2520d%2527Azur%252C%2520France%3Bss_all%3D0%3Bss_raw%3Davignon%3Bssb%3Dempty%3Bsshis%3D0%3Bssne%3DAix-en-Provence%3Bssne_untouched%3DAix-en-Provence%3Btop_ufis%3D1%26%3B&ss=Uz%C3%A8s%2C+Languedoc-Roussillon%2C+France&is_ski_area=&ssne=Avignon&ssne_untouched=Avignon&city=-1409631&checkin_year=&checkin_month=&checkout_year=&checkout_month=&group_adults=2&group_children=0&no_rooms=1&from_sf=1&ss_raw=uzes&ac_position=0&ac_langcode=fr&ac_click_type=b&dest_id=-1474231&dest_type=city&place_id_lat=44.01217&place_id_lon=4.41967&search_pageview_id=bd4030933e150131&search_selected=true&search_pageview_id=bd4030933e150131&ac_suggestion_list_length=5&ac_suggestion_theme_list_length=0&nflt=ht_id%3D204"
url_nimes = "https://www.booking.com/searchresults.fr.html?label=gen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB&sid=430afe453916ce7d5412983b211cdb2d&aid=304142&src=searchresults&error_url=https%3A%2F%2Fwww.booking.com%2Fsearchresults.fr.html%3Faid%3D304142%3Blabel%3Dgen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB%3Bsid%3D430afe453916ce7d5412983b211cdb2d%3Btmpl%3Dsearchresults%3Bac_click_type%3Db%3Bac_position%3D0%3Bcity%3D-1474231%3Bclass_interval%3D1%3Bdest_id%3D-1455068%3Bdest_type%3Dcity%3Bdtdisc%3D0%3Bfrom_sf%3D1%3Bgroup_adults%3D2%3Bgroup_children%3D0%3Biata%3DFNI%3Binac%3D0%3Bindex_postcard%3D0%3Blabel_click%3Dundef%3Bno_rooms%3D1%3Boffset%3D0%3Bpostcard%3D0%3Braw_dest_type%3Dcity%3Broom1%3DA%252CA%3Bsb_price_type%3Dtotal%3Bsearch_selected%3D1%3Bshw_aparth%3D1%3Bslp_r_match%3D0%3Bsrc%3Dsearchresults%3Bsrc_elem%3Dsb%3Bsrpvid%3De4bf30e382c50023%3Bss%3DN%25C3%25AEmes%252C%2520Languedoc-Roussillon%252C%2520France%3Bss_all%3D0%3Bss_raw%3Dnimes%3Bssb%3Dempty%3Bsshis%3D0%3Bssne%3DUz%25C3%25A8s%3Bssne_untouched%3DUz%25C3%25A8s%3Btop_ufis%3D1%26%3B&ss=N%C3%AEmes&is_ski_area=0&ssne=N%C3%AEmes&ssne_untouched=N%C3%AEmes&city=-1455068&checkin_year=&checkin_month=&checkout_year=&checkout_month=&group_adults=2&group_children=0&no_rooms=1&from_sf=1&nflt=ht_id%3D204"
url_aigues_mortes = "https://www.booking.com/searchresults.fr.html?label=gen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB&sid=430afe453916ce7d5412983b211cdb2d&aid=304142&src=searchresults&error_url=https%3A%2F%2Fwww.booking.com%2Fsearchresults.fr.html%3Faid%3D304142%3Blabel%3Dgen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB%3Bsid%3D430afe453916ce7d5412983b211cdb2d%3Btmpl%3Dsearchresults%3Bcity%3D-1455068%3Bclass_interval%3D1%3Bdest_id%3D-1455068%3Bdest_type%3Dcity%3Bdtdisc%3D0%3Bfrom_sf%3D1%3Bgroup_adults%3D2%3Bgroup_children%3D0%3Binac%3D0%3Bindex_postcard%3D0%3Blabel_click%3Dundef%3Bno_rooms%3D1%3Boffset%3D0%3Bpostcard%3D0%3Broom1%3DA%252CA%3Bsb_price_type%3Dtotal%3Bshw_aparth%3D1%3Bslp_r_match%3D0%3Bsrc%3Dsearchresults%3Bsrc_elem%3Dsb%3Bsrpvid%3Dedf730e80c7b0146%3Bss%3DN%25C3%25AEmes%3Bss_all%3D0%3Bssb%3Dempty%3Bsshis%3D0%3Bssne%3DN%25C3%25AEmes%3Bssne_untouched%3DN%25C3%25AEmes%3Btop_ufis%3D1%26%3B&ss=Aigues-Mortes%2C+Languedoc-Roussillon%2C+France&is_ski_area=&ssne=N%C3%AEmes&ssne_untouched=N%C3%AEmes&city=-1455068&checkin_year=&checkin_month=&checkout_year=&checkout_month=&group_adults=2&group_children=0&no_rooms=1&from_sf=1&ss_raw=aigues&ac_position=0&ac_langcode=fr&ac_click_type=b&dest_id=-1406800&dest_type=city&place_id_lat=43.566887&place_id_lon=4.190193&search_pageview_id=edf730e80c7b0146&search_selected=true&search_pageview_id=edf730e80c7b0146&ac_suggestion_list_length=5&ac_suggestion_theme_list_length=0&nflt=ht_id%3D204"
url_saintes_maries_de_la_mer = "https://www.booking.com/searchresults.fr.html?label=gen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB&sid=430afe453916ce7d5412983b211cdb2d&aid=304142&src=searchresults&error_url=https%3A%2F%2Fwww.booking.com%2Fsearchresults.fr.html%3Faid%3D304142%3Blabel%3Dgen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB%3Bsid%3D430afe453916ce7d5412983b211cdb2d%3Btmpl%3Dsearchresults%3Bac_click_type%3Db%3Bac_position%3D0%3Bcity%3D-1455068%3Bclass_interval%3D1%3Bdest_id%3D-1406800%3Bdest_type%3Dcity%3Bdtdisc%3D0%3Bfrom_sf%3D1%3Bgroup_adults%3D2%3Bgroup_children%3D0%3Binac%3D0%3Bindex_postcard%3D0%3Blabel_click%3Dundef%3Bnflt%3Dht_id%253D204%3Bno_rooms%3D1%3Boffset%3D0%3Bpostcard%3D0%3Braw_dest_type%3Dcity%3Broom1%3DA%252CA%3Bsb_price_type%3Dtotal%3Bsearch_selected%3D1%3Bshw_aparth%3D1%3Bslp_r_match%3D0%3Bsrc%3Dsearchresults%3Bsrpvid%3Ddd9e310baf04011b%3Bss%3DAigues-Mortes%252C%2520Languedoc-Roussillon%252C%2520France%3Bss_all%3D0%3Bss_raw%3Daigues%3Bssb%3Dempty%3Bsshis%3D0%3Bssne%3DN%25C3%25AEmes%3Bssne_untouched%3DN%25C3%25AEmes%3Btop_ufis%3D1%26%3B&ss=Les+Saintes-Maries-de-la-Mer%2C+Provence-Alpes-C%C3%B4te+d%27Azur%2C+France&is_ski_area=&ssne=Aigues-Mortes&ssne_untouched=Aigues-Mortes&city=-1406800&checkin_year=&checkin_month=&checkout_year=&checkout_month=&group_adults=2&group_children=0&no_rooms=1&from_sf=1&ss_raw=saintes&ac_position=1&ac_langcode=fr&ac_click_type=b&dest_id=-1465138&dest_type=city&place_id_lat=43.45233&place_id_lon=4.42871&search_pageview_id=dd9e310baf04011b&search_selected=true&search_pageview_id=dd9e310baf04011b&ac_suggestion_list_length=5&ac_suggestion_theme_list_length=0&nflt=ht_id%3D204"
url_collioure = "https://www.booking.com/searchresults.fr.html?label=gen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB&sid=430afe453916ce7d5412983b211cdb2d&aid=304142&src=searchresults&error_url=https%3A%2F%2Fwww.booking.com%2Fsearchresults.fr.html%3Faid%3D304142%3Blabel%3Dgen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB%3Bsid%3D430afe453916ce7d5412983b211cdb2d%3Btmpl%3Dsearchresults%3Bac_click_type%3Db%3Bac_position%3D1%3Bcity%3D-1406800%3Bclass_interval%3D1%3Bdest_id%3D-1465138%3Bdest_type%3Dcity%3Bdtdisc%3D0%3Bfrom_sf%3D1%3Bgroup_adults%3D2%3Bgroup_children%3D0%3Binac%3D0%3Bindex_postcard%3D0%3Blabel_click%3Dundef%3Bno_rooms%3D1%3Boffset%3D0%3Bpostcard%3D0%3Braw_dest_type%3Dcity%3Broom1%3DA%252CA%3Bsb_price_type%3Dtotal%3Bsearch_selected%3D1%3Bshw_aparth%3D1%3Bslp_r_match%3D0%3Bsrc%3Dsearchresults%3Bsrc_elem%3Dsb%3Bsrpvid%3D490e3112911800c0%3Bss%3DLes%2520Saintes-Maries-de-la-Mer%252C%2520Provence-Alpes-C%25C3%25B4te%2520d%2527Azur%252C%2520France%3Bss_all%3D0%3Bss_raw%3Dsaintes%3Bssb%3Dempty%3Bsshis%3D0%3Bssne%3DAigues-Mortes%3Bssne_untouched%3DAigues-Mortes%3Btop_ufis%3D1%26%3B&ss=Collioure%2C+Languedoc-Roussillon%2C+France&is_ski_area=&ssne=Les+Saintes-Maries-de-la-Mer&ssne_untouched=Les+Saintes-Maries-de-la-Mer&city=-1465138&checkin_year=&checkin_month=&checkout_year=&checkout_month=&group_adults=2&group_children=0&no_rooms=1&from_sf=1&ss_raw=collioure&ac_position=0&ac_langcode=fr&ac_click_type=b&dest_id=-1421032&dest_type=city&place_id_lat=42.52453&place_id_lon=3.082255&search_pageview_id=490e3112911800c0&search_selected=true&search_pageview_id=490e3112911800c0&ac_suggestion_list_length=5&ac_suggestion_theme_list_length=0&nflt=ht_id%3D204"
url_carcassonne = "https://www.booking.com/searchresults.fr.html?label=gen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB&sid=430afe453916ce7d5412983b211cdb2d&aid=304142&src=searchresults&error_url=https%3A%2F%2Fwww.booking.com%2Fsearchresults.fr.html%3Faid%3D304142%3Blabel%3Dgen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB%3Bsid%3D430afe453916ce7d5412983b211cdb2d%3Btmpl%3Dsearchresults%3Bac_click_type%3Db%3Bac_position%3D0%3Bcity%3D-1465138%3Bclass_interval%3D1%3Bdest_id%3D-1421032%3Bdest_type%3Dcity%3Bdtdisc%3D0%3Bfrom_sf%3D1%3Bgroup_adults%3D2%3Bgroup_children%3D0%3Binac%3D0%3Bindex_postcard%3D0%3Blabel_click%3Dundef%3Bno_rooms%3D1%3Boffset%3D0%3Bpostcard%3D0%3Braw_dest_type%3Dcity%3Broom1%3DA%252CA%3Bsb_price_type%3Dtotal%3Bsearch_selected%3D1%3Bshw_aparth%3D1%3Bslp_r_match%3D0%3Bsrc%3Dsearchresults%3Bsrc_elem%3Dsb%3Bsrpvid%3De09d31269f1d0118%3Bss%3DCollioure%252C%2520Languedoc-Roussillon%252C%2520France%3Bss_all%3D0%3Bss_raw%3Dcollioure%3Bssb%3Dempty%3Bsshis%3D0%3Bssne%3DLes%2520Saintes-Maries-de-la-Mer%3Bssne_untouched%3DLes%2520Saintes-Maries-de-la-Mer%3Btop_ufis%3D1%26%3B&ss=Carcassonne%2C+Languedoc-Roussillon%2C+France&is_ski_area=&ssne=Collioure&ssne_untouched=Collioure&city=-1421032&checkin_year=&checkin_month=&checkout_year=&checkout_month=&group_adults=2&group_children=0&no_rooms=1&from_sf=1&ss_raw=carcass&ac_position=0&ac_langcode=fr&ac_click_type=b&dest_id=-1416701&dest_type=city&iata=CCF&place_id_lat=43.21304&place_id_lon=2.34911&search_pageview_id=e09d31269f1d0118&search_selected=true&search_pageview_id=e09d31269f1d0118&ac_suggestion_list_length=5&ac_suggestion_theme_list_length=0&nflt=ht_id%3D204"
url_ariege = "https://www.booking.com/searchresults.fr.html?label=gen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB&sid=430afe453916ce7d5412983b211cdb2d&aid=304142&src=searchresults&error_url=https%3A%2F%2Fwww.booking.com%2Fsearchresults.fr.html%3Faid%3D304142%3Blabel%3Dgen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB%3Bsid%3D430afe453916ce7d5412983b211cdb2d%3Btmpl%3Dsearchresults%3Bac_click_type%3Db%3Bac_position%3D0%3Bcity%3D-1416701%3Bclass_interval%3D1%3Bdest_id%3D2507%3Bdest_type%3Dregion%3Bdtdisc%3D0%3Bfrom_sf%3D1%3Bgroup_adults%3D2%3Bgroup_children%3D0%3Binac%3D0%3Bindex_postcard%3D0%3Blabel_click%3Dundef%3Bno_rooms%3D1%3Boffset%3D0%3Bpostcard%3D0%3Braw_dest_type%3Dregion%3Broom1%3DA%252CA%3Bsb_price_type%3Dtotal%3Bsearch_selected%3D1%3Bshw_aparth%3D1%3Bslp_r_match%3D0%3Bsrc%3Dsearchresults%3Bsrc_elem%3Dsb%3Bsrpvid%3D28e1313f7d4e00a3%3Bss%3DAri%25C3%25A8ge%252C%2520France%3Bss_all%3D0%3Bss_raw%3Dariege%3Bssb%3Dempty%3Bsshis%3D0%3Bssne%3DCarcassonne%3Bssne_untouched%3DCarcassonne%3Btop_ufis%3D1%26%3B&ss=Ari%C3%A8ge&is_ski_area=0&ssne=Ari%C3%A8ge&ssne_untouched=Ari%C3%A8ge&dest_id=2507&dest_type=region&checkin_year=&checkin_month=&checkout_year=&checkout_month=&group_adults=2&group_children=0&no_rooms=1&from_sf=1&nflt=ht_id%3D204"
url_toulouse = "https://www.booking.com/searchresults.fr.html?label=gen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB&sid=430afe453916ce7d5412983b211cdb2d&aid=304142&src=searchresults&error_url=https%3A%2F%2Fwww.booking.com%2Fsearchresults.fr.html%3Faid%3D304142%3Blabel%3Dgen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB%3Bsid%3D430afe453916ce7d5412983b211cdb2d%3Btmpl%3Dsearchresults%3Bclass_interval%3D1%3Bdest_id%3D2507%3Bdest_type%3Dregion%3Bdtdisc%3D0%3Bfrom_sf%3D1%3Bgroup_adults%3D2%3Bgroup_children%3D0%3Binac%3D0%3Bindex_postcard%3D0%3Blabel_click%3Dundef%3Bno_rooms%3D1%3Boffset%3D0%3Bpostcard%3D0%3Braw_dest_type%3Dregion%3Broom1%3DA%252CA%3Bsb_price_type%3Dtotal%3Bshw_aparth%3D1%3Bslp_r_match%3D0%3Bsrc%3Dsearchresults%3Bsrc_elem%3Dsb%3Bsrpvid%3D6d1e3141f1fe0085%3Bss%3DAri%25C3%25A8ge%3Bss_all%3D0%3Bssb%3Dempty%3Bsshis%3D0%3Bssne%3DAri%25C3%25A8ge%3Bssne_untouched%3DAri%25C3%25A8ge%3Btop_ufis%3D1%26%3B&ss=Toulouse%2C+Midi-Pyr%C3%A9n%C3%A9es%2C+France&is_ski_area=&ssne=Ari%C3%A8ge&ssne_untouched=Ari%C3%A8ge&checkin_year=&checkin_month=&checkout_year=&checkout_month=&group_adults=2&group_children=0&no_rooms=1&from_sf=1&ss_raw=toulouse&ac_position=0&ac_langcode=fr&ac_click_type=b&dest_id=-1473166&dest_type=city&iata=TLS&place_id_lat=43.604645&place_id_lon=1.444213&search_pageview_id=6d1e3141f1fe0085&search_selected=true&search_pageview_id=6d1e3141f1fe0085&ac_suggestion_list_length=5&ac_suggestion_theme_list_length=0&nflt=ht_id%3D204"
url_montauban = "https://www.booking.com/searchresults.fr.html?label=gen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB&sid=430afe453916ce7d5412983b211cdb2d&aid=304142&src=searchresults&error_url=https%3A%2F%2Fwww.booking.com%2Fsearchresults.fr.html%3Faid%3D304142%3Blabel%3Dgen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB%3Bsid%3D430afe453916ce7d5412983b211cdb2d%3Btmpl%3Dsearchresults%3Bac_click_type%3Db%3Bac_position%3D0%3Bclass_interval%3D1%3Bdest_id%3D-1473166%3Bdest_type%3Dcity%3Bdtdisc%3D0%3Bfrom_sf%3D1%3Bgroup_adults%3D2%3Bgroup_children%3D0%3Biata%3DTLS%3Binac%3D0%3Bindex_postcard%3D0%3Blabel_click%3Dundef%3Bno_rooms%3D1%3Boffset%3D0%3Bpostcard%3D0%3Braw_dest_type%3Dcity%3Broom1%3DA%252CA%3Bsb_price_type%3Dtotal%3Bsearch_selected%3D1%3Bshw_aparth%3D1%3Bslp_r_match%3D0%3Bsrc%3Dsearchresults%3Bsrc_elem%3Dsb%3Bsrpvid%3D8bf03149643300d5%3Bss%3DToulouse%252C%2520Midi-Pyr%25C3%25A9n%25C3%25A9es%252C%2520France%3Bss_all%3D0%3Bss_raw%3Dtoulouse%3Bssb%3Dempty%3Bsshis%3D0%3Bssne%3DAri%25C3%25A8ge%3Bssne_untouched%3DAri%25C3%25A8ge%3Btop_ufis%3D1%26%3B&ss=Montauban%2C+Midi-Pyr%C3%A9n%C3%A9es%2C+France&is_ski_area=&ssne=Toulouse&ssne_untouched=Toulouse&city=-1473166&checkin_year=&checkin_month=&checkout_year=&checkout_month=&group_adults=2&group_children=0&no_rooms=1&from_sf=1&ss_raw=montauban&ac_position=0&ac_langcode=fr&ac_click_type=b&dest_id=-1452421&dest_type=city&place_id_lat=44.01758&place_id_lon=1.355&search_pageview_id=8bf03149643300d5&search_pageview_id=8bf03149643300d5&search_selected=true&ac_suggestion_list_length=5&ac_suggestion_theme_list_length=0&nflt=ht_id%3D204"
url_biarritz = "https://www.booking.com/searchresults.fr.html?label=gen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB&sid=430afe453916ce7d5412983b211cdb2d&aid=304142&src=searchresults&error_url=https%3A%2F%2Fwww.booking.com%2Fsearchresults.fr.html%3Faid%3D304142%3Blabel%3Dgen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB%3Bsid%3D430afe453916ce7d5412983b211cdb2d%3Btmpl%3Dsearchresults%3Bac_click_type%3Db%3Bac_position%3D0%3Bcity%3D-1473166%3Bclass_interval%3D1%3Bdest_id%3D-1452421%3Bdest_type%3Dcity%3Bdtdisc%3D0%3Bfrom_sf%3D1%3Bgroup_adults%3D2%3Bgroup_children%3D0%3Binac%3D0%3Bindex_postcard%3D0%3Blabel_click%3Dundef%3Bno_rooms%3D1%3Boffset%3D0%3Bpostcard%3D0%3Braw_dest_type%3Dcity%3Broom1%3DA%252CA%3Bsb_price_type%3Dtotal%3Bsearch_selected%3D1%3Bshw_aparth%3D1%3Bslp_r_match%3D0%3Bsrc%3Dsearchresults%3Bsrc_elem%3Dsb%3Bsrpvid%3D17c631524ebc0025%3Bss%3DMontauban%252C%2520Midi-Pyr%25C3%25A9n%25C3%25A9es%252C%2520France%3Bss_all%3D0%3Bss_raw%3Dmontauban%3Bssb%3Dempty%3Bsshis%3D0%3Bssne%3DToulouse%3Bssne_untouched%3DToulouse%3Btop_ufis%3D1%26%3B&ss=Biarritz%2C+Aquitaine%2C+France&is_ski_area=&ssne=Montauban&ssne_untouched=Montauban&city=-1452421&checkin_year=&checkin_month=&checkout_year=&checkout_month=&group_adults=2&group_children=0&no_rooms=1&from_sf=1&ss_raw=biarrit&ac_position=0&ac_langcode=fr&ac_click_type=b&dest_id=-1412526&dest_type=city&iata=BIQ&place_id_lat=43.4804&place_id_lon=-1.56066&search_pageview_id=17c631524ebc0025&search_selected=true&search_pageview_id=17c631524ebc0025&ac_suggestion_list_length=5&ac_suggestion_theme_list_length=0&nflt=ht_id%3D204"
url_bayonne = "https://www.booking.com/searchresults.fr.html?label=gen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB&sid=430afe453916ce7d5412983b211cdb2d&aid=304142&src=searchresults&error_url=https%3A%2F%2Fwww.booking.com%2Fsearchresults.fr.html%3Faid%3D304142%3Blabel%3Dgen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB%3Bsid%3D430afe453916ce7d5412983b211cdb2d%3Btmpl%3Dsearchresults%3Bac_click_type%3Db%3Bac_position%3D0%3Bcity%3D-1452421%3Bclass_interval%3D1%3Bdest_id%3D-1412526%3Bdest_type%3Dcity%3Bdtdisc%3D0%3Bfrom_sf%3D1%3Bgroup_adults%3D2%3Bgroup_children%3D0%3Biata%3DBIQ%3Binac%3D0%3Bindex_postcard%3D0%3Blabel_click%3Dundef%3Bno_rooms%3D1%3Boffset%3D0%3Bpostcard%3D0%3Braw_dest_type%3Dcity%3Broom1%3DA%252CA%3Bsb_price_type%3Dtotal%3Bsearch_selected%3D1%3Bshw_aparth%3D1%3Bslp_r_match%3D0%3Bsrc%3Dsearchresults%3Bsrc_elem%3Dsb%3Bsrpvid%3Dbdef315e4f040132%3Bss%3DBiarritz%252C%2520Aquitaine%252C%2520France%3Bss_all%3D0%3Bss_raw%3Dbiarrit%3Bssb%3Dempty%3Bsshis%3D0%3Bssne%3DMontauban%3Bssne_untouched%3DMontauban%3Btop_ufis%3D1%26%3B&ss=Bayonne%2C+Aquitaine%2C+France&is_ski_area=&ssne=Biarritz&ssne_untouched=Biarritz&city=-1412526&checkin_year=&checkin_month=&checkout_year=&checkout_month=&group_adults=2&group_children=0&no_rooms=1&from_sf=1&ss_raw=bayonne&ac_position=0&ac_langcode=fr&ac_click_type=b&dest_id=-1410844&dest_type=city&place_id_lat=43.49334&place_id_lon=-1.4751&search_pageview_id=bdef315e4f040132&search_selected=true&search_pageview_id=bdef315e4f040132&ac_suggestion_list_length=5&ac_suggestion_theme_list_length=0&nflt=ht_id%3D204"
url_la_rochelle = "https://www.booking.com/searchresults.fr.html?label=gen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB&sid=430afe453916ce7d5412983b211cdb2d&aid=304142&src=searchresults&error_url=https%3A%2F%2Fwww.booking.com%2Fsearchresults.fr.html%3Faid%3D304142%3Blabel%3Dgen173nr-1FCAEoggI46AdIDVgEaE2IAQGYAQ24ARfIAQzYAQHoAQH4AQKIAgGoAgO4Aqndno0GwAIB0gIkOTM3NmUyNmQtNjJhMC00NzI2LTliMTYtM2I1NDk1NDg1MjAy2AIF4AIB%3Bsid%3D430afe453916ce7d5412983b211cdb2d%3Btmpl%3Dsearchresults%3Bac_click_type%3Db%3Bac_position%3D0%3Bcity%3D-1412526%3Bclass_interval%3D1%3Bdest_id%3D-1410844%3Bdest_type%3Dcity%3Bdtdisc%3D0%3Bfrom_sf%3D1%3Bgroup_adults%3D2%3Bgroup_children%3D0%3Binac%3D0%3Bindex_postcard%3D0%3Blabel_click%3Dundef%3Bno_rooms%3D1%3Boffset%3D0%3Bpostcard%3D0%3Braw_dest_type%3Dcity%3Broom1%3DA%252CA%3Bsb_price_type%3Dtotal%3Bsearch_selected%3D1%3Bshw_aparth%3D1%3Bslp_r_match%3D0%3Bsrc%3Dsearchresults%3Bsrc_elem%3Dsb%3Bsrpvid%3Db5da31667363013d%3Bss%3DBayonne%252C%2520Aquitaine%252C%2520France%3Bss_all%3D0%3Bss_raw%3Dbayonne%3Bssb%3Dempty%3Bsshis%3D0%3Bssne%3DBiarritz%3Bssne_untouched%3DBiarritz%3Btop_ufis%3D1%26%3B&ss=La+Rochelle%2C+Poitou-Charentes%2C+France&is_ski_area=&ssne=Bayonne&ssne_untouched=Bayonne&city=-1410844&checkin_year=&checkin_month=&checkout_year=&checkout_month=&group_adults=2&group_children=0&no_rooms=1&from_sf=1&ss_raw=la+roc&ac_position=0&ac_langcode=fr&ac_click_type=b&dest_id=-1438604&dest_type=city&iata=LRH&place_id_lat=46.15911&place_id_lon=-1.15204&search_pageview_id=b5da31667363013d&search_selected=true&search_pageview_id=b5da31667363013d&ac_suggestion_list_length=5&ac_suggestion_theme_list_length=0&nflt=ht_id%3D204"

urls_page_city_hotels = [url_mont_saint_michel, url_saint_malo, url_bayeux, url_le_havre, url_rouen, url_paris, url_amiens, url_lille,
 url_strasbourg, url_chateau_du_haut_koenigsbourg, url_colmar, url_eguisheim, url_besançon, url_dijon, url_annecy,
 url_grenoble, url_lyon, url_gorges_du_verdon, url_bormes_les_mimosas, url_cassis, url_marseille, url_aix_en_provence,
 url_avignon, url_uzes, url_nimes, url_aigues_mortes, url_saintes_maries_de_la_mer, url_collioure, url_carcassonne,
 url_ariege, url_toulouse, url_montauban, url_biarritz, url_bayonne, url_la_rochelle]

# save the list of hotels in a txt file
F.write_txt(urls_page_city_hotels, 'data/urls.txt', mode ='w')

# save the txt file in the data lake
F.upload_file_dl('data/urls.txt', bucket_name, 'urls.txt')
