==> 4_Recommendation.ipynb <==
<p>In this last phase, we will : </p> 
<ul> 
    <li> retrieve the structured data from the database   </li> 
    <li> Vizualize recommendations </li> 
</ul>

### Table of Contents

* [1. Relational Data Storage](#section1)
* [2. Cities where the weather will be the nicest](#section2)
    * [2.1. Visualize cities with mapBox](#section21)
* [3. Discover hotels in the chosen cities](#section3)
    * [3.1. Visualize hotels with mapBox](#section31)

# access to RDS
from sqlalchemy import create_engine, text
import psycopg2

# plotting
import plotly.express as px
import plotly.io as pio
pio.renderers.default = "iframe_connected"

# 1. Relational Data Storage <a class="anchor" id="section1"></a>

dbuser = ''
dbpass = ''
dbhost = ''
dbname = ''

engine = create_engine(f"postgresql+psycopg2://{dbuser}:{dbpass}@{dbhost}/{dbname}", echo=True)

# 2. Cities where the weather will be the nicest <a class="anchor" id="section2"></a> ☀️ 😎

🗒 Comparing two cities with reference to the three criteria (temperature, humidity and percipitation_p) depends on one's perception of what could be a good weather. It depends on one's life style and especially its region climate.   
🗒  We could have used the Universal Thermal Climate Index (UTCI), however, the One Call API doesn't supply such data.   
🗒  According to this article ([Weather perception and its impact on out-of-home leisure activity participation decisions](https://www.tandfonline.com/doi/full/10.1080/21680566.2020.1733703)), temperature, percipitation and UTCI are the most important factors that may influence someone's decision of going out. That's why, we will sort our data giving three available criteria in this order :  temperature, percipitation, humidity 

n_best = int(input('How many cities would you like to get in your recommendation: '))

# get the data from the data lake
conn = engine.connect()

query = f'SELECT city.id, city.name, city.latitude, city.longitude, AVG(weather.temperature), AVG(weather.precipitation_p), AVG(weather.humidity)\
         FROM weather\
         join city\
         on city.id=weather.cid\
         GROUP BY  weather.cid, city.id \
         ORDER BY AVG(weather.temperature) desc, AVG(weather.precipitation_p) desc, AVG(weather.humidity) desc\
         LIMIT {n_best}'

stmt = text(query)
result = conn.execute(stmt)
result.fetchmany
best_cities_df = pd.DataFrame(result.fetchall(), columns=['id', 'city', 'latitude', 'longitude','avg_temperature','avg_precipitation','avg_humidity'])
best_cities_df

## 2.1. Visualize results with mapBox 📊 <a class="anchor" id="section21"></a>

fig = px.scatter_mapbox(best_cities_df, lat="latitude", lon="longitude", hover_name='city', color="avg_temperature", size =1 - best_cities_df['avg_precipitation'],
                        hover_data=['avg_temperature','avg_precipitation'], zoom=6, height=500, width=900, color_continuous_scale='bluered')

fig.update_layout(mapbox_style="open-street-map")

fig.update_layout(margin={"r":0,"t":0,"l":0,"b":0})

fig.show()

import webbrowser
import os

filename = 'file:///'+os.getcwd()+'/' + 'iframe_figures/figure_1.html'
webbrowser.open_new_tab(filename)

# 3. Discover hotels 🏩 in the chosen cities 🌇 <a class="anchor" id="section3"></a>

# this is the very simple way to interact with user without any cheking for erroneous input but
# we could use Tkinter package to make more ergonomic GUIs 
def interface():
    n_cities = int(input('Enter the number of cities to discover:\n'))
    list_cities = tuple(input("Enter the city name and press enter: \n") for _ in range(n_cities))
    criterion = input("Choose one criterion to filter hotels: \n s for star rating \n g for guests reviews \n p for price per night \n ")
    
    return list_cities, criterion

list_cities, criterion = interface()

# we suppose that there is no erroneous input
list_cities = tuple(s.lower() for s in list_cities)

filter_col = 'star'
if criterion == 'g':
    filter_col = 'rating'
elif criterion == 'p':
    filter_col = 'price'

# get the data from the data lake
conn = engine.connect()

query = f'select hotels.name, hotels.star, hotels.rating, hotels.rating_title, hotels.price, hotels.lat, hotels.lon, city.name\
          from hotels\
          join city\
          on hotels.cid = city.id\
          where lower(city.name) in {list_cities}\
          and hotels.{filter_col} is not null\
          order by hotels.{filter_col} desc'

stmt = text(query)
result = conn.execute(stmt)
result.fetchmany
best_hotels_df = pd.DataFrame(result.fetchall(), columns=['hotel_name', 'star', 'rating', 'rating_title','price','latitude','longitude', 'city_name'])
best_hotels_df

## 3.1. Visualize results with mapBox 📊 <a class="anchor" id="section31"></a>

fig = px.scatter_mapbox(best_hotels_df, lat="latitude", lon="longitude", hover_name='hotel_name', color=filter_col, size = filter_col,
                        hover_data=['city_name', 'star', 'rating', 'rating_title','price'], height=500, width=900, 
                        color_continuous_scale='oranges', zoom=7)

fig.update_layout(mapbox_style="open-street-map")

fig.update_layout(margin={"r":0,"t":0,"l":0,"b":0})

fig.show()

import webbrowser
import os

filename = 'file:///'+os.getcwd()+'/' + 'iframe_figures/figure_2.html'
webbrowser.open_new_tab(filename)
